fix: honor wallet-frontend .env file
This allows for different configurations that are not under version
control. For instance, on a Digg computer we normally run Docker inside
WSL but access the wallet frontend from a web browser directly in
Windows. On the Windows host we are not able to modify the hosts mapping
(etc/hosts). Instead we access the docker services through port
forwarding on localhost. On a standard setup where we are allowed to
edit the hosts file, we want to access the services using their host
names.

In order to support both use cases, we let the user configure their own
environment. However, if the .env file does not exist, we preserve the
current behavior of copying the template file.

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>