perf: reduce files included in build of custom-verifier
By limiting the amount of files we include in the Docker build we avoid
rebuilding the image due to irrelevant changes. For instance, we
generally do not need to rebuild the Docker image when something has
changed in the ./docker directory.

This reduces the build time of the custom-verifier Docker image.

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>