test: add ability to run test suite once
On a development machine, `npm run test` will enter watch mode and
repeatedly rerun the test suite on any file change. This is generally
useful, but sometimes we want to run the test suite only once and then
exit - for instance, when used to guard pushing changes to a remote
repository. To suppport this use case, and others, we introduce the `npm
run test:once` command:

	npm run test:once && git push

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>