feat: use GitHub version of Megalinter
By using the same version of Megalinter as we do on GitHub we increase
the likelihood of similar linting results. For instance, the Megalinter
version determines the Gitleaks version. Gitleaks seems very sensitive
to version differences. By keeping the versions in sync we ensure that
the rules we develop work the same when running on GitHub.

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>