fix: remove ineffective Gitleaks rule
The `[[allowlists]]` syntax was introduced in version 8.25.0 of
Gitleaks, but we are using version 8.20.1 on GitHub and locally. Hence
the allow list for the the HAProxy PEM file is ineffective. Since we
don't get any failures regarding this file, we can safely remove the
allow list.

See: https://github.com/gitleaks/gitleaks?tab=readme-ov-file
Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>