test: add simple test runner for Gitleaks rules
This makes it simple to verify that you did not break anything when
updating our custom Gitleaks rules. The test runner is based on the
approval testing methodology where we run a program, e.g. our Gitleaks
rules check, and compare the output to a stored and approved copy of a
previous run. If the output is identical, the test passes. Otherwise the
test fails.

How to run:

	.gitleaks-test/test-runner.sh .gitleaks-test/tests/suite.txt logs/gitleaks-test

See: https://approvaltests.com/
Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>