feat: allow customization of commit lint baseline
Normally we want to compare against the 'main' branch when running the
commit lint. However, the mainline is not always called 'main'. For
instance, when doing development on a forked version of the repository,
the 'main' branch is probably on the origin and the mainline of the
forked repo is called something else. By using an environment variable
we allow for comparing with the mainline of the forked reposistory.

Assuming you have a remote named 'my-fork' you can run the commit lint
using the 'main' branch of your fork using the command below:

	 env COMMIT_COMPARE_TO_BRANCH=my-fork/main ./development/code_quality.sh

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>