fix: copy proxy certificate together with node package files
In Digg's build pipeline we use `buildah` to build the container image.
It turns out that buildah does not like wildcard copies that do not
match any file. When this happens the build fails. In the build pipeline
we never side load the certificate since it is bundled with the
container image. However, we still want to support this use case in
order to allow side loading when running locally.

By merging the copy operation for the node package files with the one
for the the optional certificate we ensure that there is always at least
some file to copy, e.g. the package.json file, and avoid build failure.

Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>