feat: configure stricter linting
This change set configures MegaLinter to be more strict. Specifically,
with the updated configuration we will fail the build in any of the
following situations:

 1. When there are any formatting issues detected
 2. When there are any automatic fixes suggested by any of the linters

Without these changes, formatting issues and automatic fixes are
detected and the source files are updated. However, the updates are not
committed - and worse, the issues are not even reported.

See: https://megalinter.io/latest/config-variables/
Signed-off-by: Eric Thelin <extern.eric.thelin@digg.se>